A lane of five steps jumped off the beat once an hour
The global step counter was masked with 0x7FFF. 32768 is a power of two, so
8-, 16- and 32-step lanes wrapped cleanly - which is every lane in the shipped
guide, and every lane anyone had tested. A 5-, 6- or 12-step lane jumped by
(32768 % steps) every time the counter rolled: once every 66 minutes at
124 bpm.

A pattern silently moving off the beat, once an hour, in the middle of a set.
It is the worst kind of bug for this device - too rare to reproduce on demand,
too specific to notice unless you happened to be playing an odd-length lane
for an hour, and indistinguishable from "the ESP32 just drifts".

There is no reason for the mask. The counter is 32 bits and advances at the
step rate, so unmasked it rolls after about eight years of continuous playing,
and a lane finds its own step with `% lane->steps` either way.

The check asserts the property directly rather than the fix: consecutive steps
must advance by exactly one, modulo the lane length, across the roll - for
every lane length from 2 to 32. It also asserts that the OLD mask IS a
counterexample at 5 steps, so the test is known to be capable of detecting the
thing it was written for. That is the same standard as the CASET check, where
the naive implementation is right for the common case and wrong for the one
that matters.

Verified on hardware, which is also the bench check the playhead design asked
for and had not had: a 16-step kick against a 12-step hat. They coincide at
13723, 14207 and 14691 ms and then visibly separate, the kick holding 484 ms
throughout. A global position would have kept them locked forever; per-lane
modulo lets them drift, which is what a polyrhythm is.

+++ firmware/main/editor.c
@@ -335,7 +335,7 @@ static int playhead_offset(int line_off, const char *lbuf, int at, int len)
     }
     /* The global step is not the lane's step the moment one lane is not the
      * same length as another - which is the point of having lanes. */
-    const int off = seq_pattern_offset(lbuf + a, seq_position() % l->steps,
+    const int off = seq_pattern_offset(lbuf + a, (int)(seq_position() % (uint32_t)l->steps),
                                        SEQ_MAX_STEPS);
     return (off < 0) ? -1 : line_off + a + off;
 }
